requirements.txt
streamlit
crewai
google-generativeai
python-dotenv
langchain
langchain-community
crewai-tools
langchain-google-genai
plotly
textblob
streamlit-extras